==> 01_makerlabs_data.ipynb <==
# 01 Makerlabs data

import os
from makerlabs import fablabs_io
from makerlabs import hackerspaces_org
from makerlabs import makerspaces_make_co
from makerlabs import repaircafe_org

orstoken = ""
with open('orstoken.txt', 'r') as f:
    orstoken = f.readline()
    orstoken = orstoken.rstrip()

open_cage_api_key=orstoken

diybios = diybio_org.get_labs(format="pandas", open_cage_api_key=open_cage_api_key)

fablabs = fablabs_io.get_labs(format="pandas", open_cage_api_key=open_cage_api_key)

hackerspaces = hackerspaces_org.get_labs(format="pandas", open_cage_api_key=open_cage_api_key)

makerspaces = makerspaces_make_co.get_labs(format="pandas", open_cage_api_key=open_cage_api_key)

repaircafes = repaircafe_org.get_labs(format="pandas", open_cage_api_key=open_cage_api_key)

fablabs.to_csv("data/fablabs.csv")
hackerspaces.to_csv("data/hackerspaces.csv")
makerspaces.to_csv("data/makerspaces.csv")
repaircafes.to_csv("data/repaircafes.csv")
